Started automated test files
