+++ .vscode/settings.json
@@ -38,6 +38,10 @@
     "raspberry-pi-pico.ninjaPath": "${HOME}/.pico-sdk/ninja/v1.12.1/ninja",
     "raspberry-pi-pico.python3Path": "python3",
     "files.associations": {
-        "stdlib.h": "c"
+        "stdlib.h": "c",
+        "functions.h": "c",
+        "freertos.h": "c",
+        "cyw43_arch.h": "c",
+        "multicore.h": "c"
     }
 }
